+++ src/stack_utils.c
@@ -6,7 +6,7 @@
 /*   By: lhenriqu <lhenriqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/29 11:19:47 by lhenriqu          #+#    #+#             */
-/*   Updated: 2025/01/30 12:50:41 by lhenriqu         ###   ########.fr       */
+/*   Updated: 2025/01/30 15:53:45 by lhenriqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ t_stack	*get_min_greater_than(t_stack *stk_a, int position)
 	return (min);
 }
 
-t_stack	*get_min_node(t_stack *stack, int n)
+t_stack	*get_min_node(t_stack *stack, long n)
 {
 	t_stack	*min;
 
@@ -60,7 +60,7 @@ t_stack	*get_max_node(t_stack *stack)
 t_stack	*get_min_position(t_stack *stack, int max)
 {
 	if (max == 0)
-		return (get_min_node(stack, INT_MIN));
+		return (get_min_node(stack, LONG_MIN));
 	while (stack)
 	{
 		if (stack->sorted_position <= max)
